refactor region finding methods
