+++ optim_esm_tools/__init__.py
@@ -5,6 +5,7 @@ __author__ = 'Joran R. Angevaare'
 from . import utils
 from . import config
 from . import analyze
+from . import region_finding
 from . import synda_files
 from . import _test_utils
 from . import plotting
